+++ examples/01_plotting/plot_visualize_megatrawls_netmats.py
@@ -24,14 +24,11 @@ partial_correlation = netmats.correlation_matrices
 ################################################################################
 # Visualization
 # -------------
-# Import matplotlib and nilearn plotting modules to use its utilities for
-# plotting correlation matrices
-import matplotlib.pyplot as plt
+# Import nilearn plotting modules to use its utilities for plotting
+# correlation matrices
 from nilearn import plotting
 
-title = "Partial correlation matrices of d=300'"
-plt.figure()
-plt.imshow(partial_correlation, interpolation="nearest", cmap=plotting.cm.bwr)
-plt.colorbar()
-plt.title(title)
-plt.show()
+title = "Partial correlation matrices\n for d=300"
+display = plotting.plot_matrix(partial_correlation, colorbar=True,
+                               title=title)
+plotting.show()