+++ examples/03_connectivity/plot_inverse_covariance_connectome.py
@@ -6,7 +6,7 @@ This example constructs a functional connectome using the sparse inverse
 covariance.
 
 We use the `MSDL atlas
-<https://team.inria.fr/parietal/research/spatial_patterns/spatial-patterns-in-resting-state/>`_
+<https://team.inria.fr/parietal/18-2/spatial_patterns/spatial-patterns-in-resting-state/>`_
 of functional regions in rest, and the
 :class:`nilearn.input_data.NiftiMapsMasker` to extract time series.
 
@@ -59,23 +59,17 @@ estimator.fit(time_series)
 ##############################################################################
 # Display the connectome matrix
 # ------------------------------
-from matplotlib import pyplot as plt
-
+from nilearn import plotting
 # Display the covariance
-plt.figure(figsize=(10, 10))
 
 # The covariance can be found at estimator.covariance_
-plt.imshow(estimator.covariance_, interpolation="nearest",
-           vmax=1, vmin=-1, cmap=plt.cm.RdBu_r)
-# And display the labels
-x_ticks = plt.xticks(range(len(labels)), labels, rotation=90)
-y_ticks = plt.yticks(range(len(labels)), labels)
-plt.title('Covariance')
+plotting.plot_matrix(estimator.covariance_, labels=labels,
+                     figure=(9, 7), vmax=1, vmin=-1,
+                    title='Covariance')
 
 ##############################################################################
 # And now display the corresponding graph
 # ----------------------------------------
-from nilearn import plotting
 coords = atlas.region_coords
 
 plotting.plot_connectome(estimator.covariance_, coords,
@@ -86,13 +80,9 @@ plotting.plot_connectome(estimator.covariance_, coords,
 # Display the sparse inverse covariance
 # --------------------------------------
 # we negate it to get partial correlations
-plt.figure(figsize=(10, 10))
-plt.imshow(-estimator.precision_, interpolation="nearest",
-           vmax=1, vmin=-1, cmap=plt.cm.RdBu_r)
-# And display the labels
-x_ticks = plt.xticks(range(len(labels)), labels, rotation=90)
-y_ticks = plt.yticks(range(len(labels)), labels)
-plt.title('Sparse inverse covariance')
+plotting.plot_matrix(-estimator.precision_, labels=labels,
+                     figure=(9, 7), vmax=1, vmin=-1,
+                     title='Sparse inverse covariance')
 
 ##############################################################################
 # And now display the corresponding graph