+++ examples/03_connectivity/plot_probabilistic_atlas_extraction.py
@@ -6,7 +6,7 @@ This example extracts the signal on regions defined via a probabilistic
 atlas, to construct a functional connectome.
 
 We use the `MSDL atlas
-<https://team.inria.fr/parietal/research/spatial_patterns/spatial-patterns-in-resting-state/>`_
+<https://team.inria.fr/parietal/18-2/spatial_patterns/spatial-patterns-in-resting-state/>`_
 of functional regions in rest.
 
 The key to extract signals is to use the
@@ -59,17 +59,11 @@ correlation_matrix = correlation_measure.fit_transform([time_series])[0]
 
 # Display the correlation matrix
 import numpy as np
-from matplotlib import pyplot as plt
-plt.figure(figsize=(10, 10))
+from nilearn import plotting
 # Mask out the major diagonal
 np.fill_diagonal(correlation_matrix, 0)
-plt.imshow(correlation_matrix, interpolation="nearest", cmap="RdBu_r",
-           vmax=0.8, vmin=-0.8)
-plt.colorbar()
-# And display the labels
-x_ticks = plt.xticks(range(len(labels)), labels, rotation=90)
-y_ticks = plt.yticks(range(len(labels)), labels)
-
+plotting.plot_matrix(correlation_matrix, labels=labels, colorbar=True,
+                     vmax=0.8, vmin=-0.8)
 ############################################################################
 # And now display the corresponding graph
 # ----------------------------------------